+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 Added support for the :let keyword in the command interface.
 
+Changed `defcommand` to create a variadic function.
+
 # v0.3 -- 11 Aug 2022
 
 Fix logic for default command summary to just be first line of docstring.