+++ CHANGES.md
@@ -4,6 +4,10 @@ Added support for the :let keyword in the command interface.
 
 Changed `defcommand` to create a variadic function.
 
+Added checks that options and argument symbols are unique; now all parsed options
+and positional arguments are provided in a single `:options` map, which simplifies
+testing.
+
 # v0.3 -- 11 Aug 2022
 
 Fix logic for default command summary to just be first line of docstring.